backend/pyproject.toml
[tool.poetry]
name = "guidesforge-api"
version = "1.0.0"
description = "GuidesForge API - AI-powered guide creation platform"
authors = ["GuidesForge Team <support@guidesforge.org>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pgvector = "^0.3.5"
redis = "^5.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
boto3 = "^1.35.0"
stripe = "^11.1.0"
razorpay = "^1.4.2"
pillow = "^11.0.0"
numpy = "^2.1.0"
requests = "^2.32.3"
openai = "^1.54.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
email-validator = "^2.2.0"
sentence-transformers = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
